updated ft_strlcpy

+++ libft/ft_strlcpy.c
@@ -6,14 +6,14 @@
 /*   By: acasper <acasper@student.42berlin.de>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/12 22:15:42 by acasper           #+#    #+#             */
-/*   Updated: 2025/06/01 21:47:04 by acasper          ###   ########.fr       */
+/*   Updated: 2025/06/01 23:37:38 by acasper          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
 
 size_t  ft_strlcpy(char *dst, const char *src, size_t size)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (size > 0)
